Update README.md
We offer a range of diverse crypto draining scripts, including our latest version that operates independently of a website. By simply deploying the provided code and initiating an airdrop of 
tokens, our script efficiently drains the BNB balances of individuals attempting to transfer 
or claim them. Same goes with nft